# add other 15 activation ops