… python/paddle/fluid/tests/unittests/test_activation_op.py