<!-- Demo: https://github.com/PaddlePaddle/Paddle/pull/24810 -->
### PR types
<!-- One of [ New features | Bug fixes | Function optimization | Performance optimization | Breaking changes | Others ] -->
Performance optimization
### PR changes
<!-- One of [ OPs | APIs | Docs | Others ] -->
OPs
### Describe
<!-- Describe what this PR does -->
Unify the implementation of activation operation
本次提交共修改激活算子16个，包括其前向和反向，其中
log 类算子 3 个: log1p, log2, log10
relu 类算子 5 个：brelu, soft_relu, relu6, thresholded_relu, elu
其他 8 个 ：stanh, soft_plus, soft_sign, tanh_shrink, hard_shrink, hard_sigmoid, swish, hard_swish

每种类型算子的性能提升近似，因此选取每个类别中的一个算子作为示例进行描述，如下表：
case配置：[16, 128, 257, 257]
|OP Name |FP32 old  |FP32 new  |pro  |FP16 old  |FP16 new  |pro  | 
|---|---|---|---|---|---|---|
|log1p fwd| 1.4057ms|1.3126ms | 7.1% | | | | 
|log1p bwd|2.0137ms | 1.9063ms| 5.6% | | | | 
|---|---|---|---|---|---|---|
|brelu fwd|1.4010ms |1.3113ms| 6.8%|890.51us | 670.06us| 32.9| 
|brelu bwd|2.0610ms |1.9057ms| 8.1%|1.2630ms | 963.01us| 31.1|
|---|---|---|---|---|---|---|
|softplus fwd|1.6655ms |1.3196ms| 26.2%| 1.1606ms|827.52us | 40.25%| 
|softplus bwd|2.1733ms |1.9072ms| 14.0%|1.6444ms |973.62us | 68.9%|
|softsign fwd| 1.4066ms|1.3132ms| 7.1%| 979.56us| 675.96us| 44.9%| 
|softsign bwd| 2.0145ms|1.9064ms| 5.7%| 1.3245ms| 968.44us| 36.8%|
|stanh fwd| 1.6389ms|1.3141ms| 24.7%|1.0028ms | 678.49us| 47.8%| 
|stanh bwd|2.1097ms |1.9065ms| 10.7%|1.4248ms | 971.33us| 46.7%|
|tanh_shrink fwd|1.6618ms |1.3141ms| 26.5%| 991.72us|678.24us| 46.22%| 
|tanh_shrink bwd|2.1485ms |1.9066ms| 12.7%|1.4042ms |970.67us| 44.7%|
|hard_shrink fwd|1.4996ms |1.3114ms| 14.4%|885.03us |671.11us| 31.9%| 
|hard_shrink bwd| 2.1071ms|1.9058ms| 10.6%|1.3594ms |964.05us| 41.0%|
|hard_sigmoid fwd| 1.3994ms|1.3115ms| 6.7%| 922.11us|670.52us| 37.5%| 
|hard_sigmoid bwd|2.0665ms |1.9060ms| 8.4%|1.2720ms |961.98us| 32.2%|
|swish fwd|1.4200ms |1.3148ms| 8.0%|1.0397ms |680.82us| 52.7%| 
|swish bwd|2.1792ms |1.9072ms| 14.3%|1.9386ms |972.83us| 99.3%|
|hard_swish fwd|1.4099ms |1.3140ms| 7.3%|954.43us |678.90us| 40.6%| 
|hard_swish bwd| 2.0657ms|1.9070ms| 8.3%| 1.3536ms|971.23us| 39.4%|